무결성 문제 해결 clear
