+++ sw_mega/src/test/java/com/pcwk/ehr/board/FbCommentDaoTest.java
@@ -1,50 +1,66 @@
 package com.pcwk.ehr.board;
 
+import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
+
+import java.util.Date;
 
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Disabled;
+import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.context.ApplicationContext;
 import org.springframework.test.context.ContextConfiguration;
 import org.springframework.test.context.junit.jupiter.SpringExtension;
 
-import com.pcwk.ehr.board.mapper.FreeBoardMapper;
+import com.pcwk.ehr.board.domain.FreeBoardCommentDTO;
+import com.pcwk.ehr.board.mapper.FbCommentMapper;
+
+
 
 @ExtendWith(SpringExtension.class)
-@ContextConfiguration(locations = { "file:src/main/webapp/WEB-INF/spring/root-context.xml",
-		"file:src/main/webapp/WEB-INF/spring/appServlet/servlet-context.xml" })
-class FreeBoardTest {
+@ContextConfiguration(locations = {
+    "file:src/main/webapp/WEB-INF/spring/root-context.xml",
+    "file:src/main/webapp/WEB-INF/spring/appServlet/servlet-context.xml"
+})
 
-	Logger log = LogManager.getLogger(getClass());
+class FbCommentDaoTest {
+    final Logger log = LogManager.getLogger(getClass());
 
-	@Autowired
-	FreeBoardMapper freeBoardMapper;
+    @Autowired
+    ApplicationContext context;
 
+    @Autowired
+    FbCommentMapper fbCommentMapper;
+    
+    
 	@BeforeEach
-	void setUp() {
+	void setUp() throws Exception {
 		log.debug("┌────────────────────────────┐");
 		log.debug("│ setUp()                    │");
 		log.debug("└────────────────────────────┘");
 	}
 
+ 
+       
+	
 	@AfterEach
-	void tearDown() {
+	void tearDown() throws Exception {
 		log.debug("┌────────────────────────────┐");
 		log.debug("│ tearDown()                 │");
 		log.debug("└────────────────────────────┘");
-
 	}
 
-	// @Disabled
-	@Test
-	void beans() {
-
-		assertNotNull(freeBoardMapper);
-		log.debug(freeBoardMapper);
+	//@Disabled
+    @Test
+    void beans() {
+        assertNotNull(context);
+        assertNotNull(context);
 	}
 
 }
