+++ sw_mega/src/test/java/com/pcwk/ehr/board/FreeBoardDaoTest.java
@@ -1,66 +1,50 @@
 package com.pcwk.ehr.board;
 
-import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
-
-import java.util.Date;
 
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.Disabled;
-import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.context.ApplicationContext;
 import org.springframework.test.context.ContextConfiguration;
 import org.springframework.test.context.junit.jupiter.SpringExtension;
 
-import com.pcwk.ehr.board.domain.FreeBoardCommentDTO;
-import com.pcwk.ehr.board.mapper.FbCommentMapper;
-
-
+import com.pcwk.ehr.board.mapper.FreeBoardMapper;
 
 @ExtendWith(SpringExtension.class)
-@ContextConfiguration(locations = {
-    "file:src/main/webapp/WEB-INF/spring/root-context.xml",
-    "file:src/main/webapp/WEB-INF/spring/appServlet/servlet-context.xml"
-})
+@ContextConfiguration(locations = { "file:src/main/webapp/WEB-INF/spring/root-context.xml",
+		"file:src/main/webapp/WEB-INF/spring/appServlet/servlet-context.xml" })
+class FreeBoardDaoTest {
 
-class Fb_CommentTest {
-    final Logger log = LogManager.getLogger(getClass());
+	Logger log = LogManager.getLogger(getClass());
 
-    @Autowired
-    ApplicationContext context;
+	@Autowired
+	FreeBoardMapper freeBoardMapper;
 
-    @Autowired
-    FbCommentMapper fbCommentMapper;
-    
-    
 	@BeforeEach
-	void setUp() throws Exception {
+	void setUp() {
 		log.debug("┌────────────────────────────┐");
 		log.debug("│ setUp()                    │");
 		log.debug("└────────────────────────────┘");
 	}
 
- 
-       
-	
 	@AfterEach
-	void tearDown() throws Exception {
+	void tearDown() {
 		log.debug("┌────────────────────────────┐");
 		log.debug("│ tearDown()                 │");
 		log.debug("└────────────────────────────┘");
+
 	}
 
-	//@Disabled
-    @Test
-    void beans() {
-        assertNotNull(context);
-        assertNotNull(context);
+	// @Disabled
+	@Test
+	void beans() {
+
+		assertNotNull(freeBoardMapper);
+		log.debug(freeBoardMapper);
 	}
 
 }
